:sparkles: feat: Update Professional With Toast and Modal

+++ src/app/modules/maintenance/pages/professionals-table-page/professionals-table-page.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Page } from '../../../../core/models/page';
 import { Professional } from '../../../../core/models/professional';
 import { ProfessionalService } from '../../../../core/services/professional.service';
+import { ModalComponent } from '../../../../shared/components/modal/modal.component';
+import { ToastService } from '../../../../core/services/toast.service';
 
 @Component({
   selector: 'app-professionals-table-page',
@@ -12,8 +14,13 @@ export class ProfessionalsTablePageComponent implements OnInit{
   ProfessionalPage: Page<Professional> = {} as Page<Professional>
   filter: string = ""
   page: number = 1;
+  selectedProfessional: Professional = {} as Professional
 
-  constructor(private professionalService: ProfessionalService){}
+  constructor(
+    private professionalService: ProfessionalService,
+    private toastService: ToastService
+
+  ){}
 
   ngOnInit(): void {
     this.loadProfessionals()
@@ -24,9 +31,8 @@ export class ProfessionalsTablePageComponent implements OnInit{
       next: response => {
         this.ProfessionalPage.content = response.body
         this.ProfessionalPage.numberofElements = parseInt(response.headers.get('X-Total-Count')  || '0')
-        console.log(this.ProfessionalPage.content)
       },
-      error: () => console.log("Erro ao procurar profissionais")
+      error: () => this.toastService.show("Erro ao carregar os profissionais", {classname: 'bg-danger text-light'})
     })
   }
 
@@ -38,13 +44,18 @@ export class ProfessionalsTablePageComponent implements OnInit{
     this.loadProfessionals()
   }
 
-  removeProfesional(professional: Professional){
-    this.professionalService.deleteProfessional(professional).subscribe({
-      next: () => {
-        console.log("Profissional removido com sucesso")
-        this.loadProfessionals()
-      },
-      error: () => console.log("Erro ao excluir o profissional")
+  removeProfesional(professional: Professional, modalConfirm: ModalComponent){
+    this.selectedProfessional = professional
+    modalConfirm.open().then(confirm => {
+      if(confirm){
+        this.professionalService.deleteProfessional(professional).subscribe({
+          next: () => {
+            this.toastService.show("Profissional removido com sucesso", {classname: 'bg-success text-light'})
+            this.loadProfessionals()
+          },
+          error: () => this.toastService.show("Erro ao excluir o profissional", {classname: 'bg-danger text-light'})
+        })
+      }
     })
   }
